datus-storage-postgresql/pyproject.toml
[project]
name = "datus-storage-postgresql"
version = "0.1.0"
description = "PostgreSQL RDB and pgvector backend adapters for datus-agent"
requires-python = ">=3.12"
dependencies = [
    "datus-agent",
    "psycopg[binary]>=3.1",
    "psycopg-pool>=3.1",
    "pgvector>=0.3.0",
    "pyarrow",
    "pandas",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "testcontainers[postgres]>=4.0",
]

[project.entry-points."datus.storage.rdb"]
postgresql = "rdb:register"

[project.entry-points."datus.storage.vector"]
postgresql = "vector:register"

[build-system]
requires = ["setuptools>=80.9.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
